+++ src/Presentation/SmartStore.Web/Administration/Validators/Common/GenericAttributeValidator.cs
@@ -1,0 +1,14 @@
+﻿using FluentValidation;
+using SmartStore.Admin.Models.Common;
+using SmartStore.Services.Localization;
+
+namespace SmartStore.Admin.Validators.Localization
+{
+    public class GenericAttributeValidator : AbstractValidator<GenericAttributeModel>
+    {
+        public GenericAttributeValidator(ILocalizationService localizationService)
+        {
+            RuleFor(x => x.Key).NotEmpty().WithMessage(localizationService.GetResource("Admin.Common.GenericAttributes.Fields.Name.Required"));
+        }
+    }
+}